+++ src/aks-preview/azext_aks_preview/_help.py
@@ -586,6 +586,12 @@ helps['aks create'] = f"""
         - name: --enable-app-routing
           type: bool
           short-summary: Enable Application Routing addon.
+        - name: --enable-secure-boot
+          type: bool
+          short-summary: Enable Secure Boot on all node pools in the cluster.
+        - name: --enable-vtpm
+          type: bool
+          short-summary: Enable vTPM on all node pools in the cluster.
     examples:
         - name: Create a Kubernetes cluster with an existing SSH public key.
           text: az aks create -g MyResourceGroup -n MyManagedCluster --ssh-key-value /path/to/publickey
@@ -1133,6 +1139,18 @@ helps['aks update'] = """
         - name: --node-provisioning-mode
           type: string
           short-summary: Set the node provisioning mode of the cluster. Valid values are "Auto" and "Manual". For more information on "Auto" mode see aka.ms/aks/nap.
+        - name: --enable-secure-boot
+          type: bool
+          short-summary: Enable Secure Boot.
+        - name: --disble-secure-boot
+          type: bool
+          short-summary: Disable Secure Boot.
+        - name: --enable-vtpm
+          type: bool
+          short-summary: Enable vTPM.
+        - name: --disble-vtpm
+          type: bool
+          short-summary: Disable vTPM.
     examples:
       - name: Reconcile the cluster back to its current state.
         text: az aks update -g MyResourceGroup -n MyManagedCluster