+++ src/aks-preview/azext_aks_preview/custom.py
@@ -613,6 +613,9 @@ def aks_create(
     storage_pool_sku=None,
     storage_pool_option=None,
     node_provisioning_mode=None,
+    # trusted launch
+    enable_secure_boot=False,
+    enable_vtpm=False,
 ):
     # DO NOT MOVE: get all the original parameters and save them as a dictionary
     raw_parameters = locals()
@@ -799,6 +802,11 @@ def aks_update(
     storage_pool_option=None,
     azure_container_storage_nodepools=None,
     node_provisioning_mode=None,
+    # trusted launch
+    enable_secure_boot=False,
+    disable_secure_boot=False,
+    enable_vtpm=False,
+    disable_vtpm=False,
 ):
     # DO NOT MOVE: get all the original parameters and save them as a dictionary
     raw_parameters = locals()